+++ src/main/java/com/linkode/api_server/service/GithubIssueService.java
@@ -1,6 +1,5 @@
 package com.linkode.api_server.service;
 
-import com.linkode.api_server.common.exception.GithubIssueException;
 import com.linkode.api_server.common.exception.MemberStudyroomException;
 import com.linkode.api_server.common.exception.StudyroomException;
 import com.linkode.api_server.domain.GithubIssue;
@@ -8,7 +7,6 @@ import com.linkode.api_server.domain.Studyroom;
 import com.linkode.api_server.domain.base.BaseStatus;
 import com.linkode.api_server.dto.gitHubIssue.GithubIssueListResponse;
 import com.linkode.api_server.dto.gitHubIssue.GithubIssueResponse;
-import com.linkode.api_server.dto.studyroom.UploadDataResponse;
 import com.linkode.api_server.repository.githubIssue.GithubIssueDSL;
 import com.linkode.api_server.repository.githubIssue.GithubIssueRepository;
 import com.linkode.api_server.repository.MemberstudyroomRepository;
@@ -16,19 +14,12 @@ import com.linkode.api_server.repository.StudyroomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.client.RestClientException;
-import org.springframework.web.client.RestTemplate;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+
 
 import static com.linkode.api_server.common.response.status.BaseExceptionResponseStatus.*;
 
@@ -42,10 +33,6 @@ public class GithubIssueService {
     private final GithubIssueDSL githubIssueDSL;
     private final StudyroomRepository studyroomRepository;
     private final MemberstudyroomRepository memberstudyroomRepository;
-    private final RestTemplate restTemplate;
-
-    @Value("${SOCKET_SERVER_URL}")
-    private String SOCKET_SERVER_URL;
 
     @Transactional
     public GithubIssueResponse saveGithubIssue(Long studyroomId, String payload) {
@@ -95,27 +82,5 @@ public class GithubIssueService {
         return GithubIssueListResponse.builder().IssueList(issues).build();
     }
 
-    /** 업로드 응답을 메세지로 브로드캐스트 */
-    public void broadcastGithubIssue(long studyroomId, GithubIssueResponse issue) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
 
-        Map<String, Object> body = new HashMap<>();
-        body.put("studyroomId", studyroomId);
-        body.put("event", "issueUploaded");
-        body.put("data", issue);
-
-        HttpEntity<Map<String, Object>> request = new HttpEntity<>(body, headers);
-
-        try {
-            ResponseEntity<String> responseEntity = restTemplate.postForEntity(SOCKET_SERVER_URL, request, String.class);
-            if (responseEntity.getStatusCode().is2xxSuccessful()) {
-                log.info("Broadcast sent successfully");
-            } else {
-                log.error("Failed to send broadcast. Status code: " + responseEntity.getStatusCodeValue());
-            }
-        } catch (RestClientException e) {
-            log.error("Error sending broadcast: " + e.getMessage());
-        }
-    }
 }
